+++ packages/client/src/lib/drizzle/schema/llmLocalSessions.ts
@@ -1,0 +1,30 @@
+import { sortByKey } from "@/lib/utils";
+import { relations, sql } from "drizzle-orm";
+import { blob, integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
+import { llmCompletions } from "./llmCompletions";
+
+export const llmLocalSessions = sqliteTable(
+  "llm_local_sessions",
+  sortByKey({
+    id: integer("id").primaryKey(),
+    internalId: text("internal_id").notNull(),
+    modelPath: text("model_path").notNull(),
+
+    /**
+     * NOTE: Accepts `string` as a type.
+     */
+    modelHash: blob("model_hash").notNull(),
+
+    contextSize: integer("context_size").notNull(),
+    createdAt: integer("created_at", { mode: "timestamp" })
+      .notNull()
+      .default(sql`strftime('%s', 'now')`),
+  }),
+);
+
+export const llmLocalSessionRelatiosn = relations(
+  llmLocalSessions,
+  ({ many }) => ({
+    completions: many(llmCompletions),
+  }),
+);