+++ packages/client/src/lib/inference/RemoteLlmDriver.ts
@@ -1,11 +1,15 @@
 import * as api from "@/lib/api";
+import { eq } from "drizzle-orm";
 import { toMilliseconds } from "duration-fns";
 import { Ref, ref } from "vue";
-import { mergeAbortSignals, timeoutSignal } from "../utils";
+import { d } from "../drizzle";
+import { LatestSession } from "../storage/llm";
+import { mergeAbortSignals, omit, timeoutSignal } from "../utils";
 import {
   BaseLlmDriver,
   CompletionOptions,
   CompletionProgressEventPayload,
+  CompletionResult,
   DecodeProgressEventPayload,
   LlmGrammarLang,
 } from "./BaseLlmDriver";
@@ -19,18 +23,71 @@ export type RemoteLlmDriverConfig = {
 export class RemoteLlmDriver implements BaseLlmDriver {
   readonly busy = ref(false);
 
+  /**
+   * Create a new RemoteLlmDriver instance.
+   */
   static async create(
     config: RemoteLlmDriverConfig,
+    latestSession: Ref<LatestSession | null>,
     jwt: Readonly<Ref<string | null>>,
-    sessionId: Ref<string | null>,
   ) {
+    let databaseSession;
+    outer: if (latestSession.value) {
+      if (latestSession.value.driver !== "remote") {
+        console.warn(`Latest session driver mismatch`, { latestSession });
+        break outer;
+      }
+
+      databaseSession = await d.db.query.llmRemoteSessions.findFirst({
+        where: eq(d.llmRemoteSessions.id, latestSession.value.id),
+      });
+
+      if (!databaseSession) {
+        console.warn(`Latest session not found in DB`, { latestSession });
+        latestSession.value = null;
+        break outer;
+      }
+
+      if (databaseSession.baseUrl !== config.baseUrl) {
+        console.warn(`Config base URL mismatch`, {
+          databaseSessionId: databaseSession.id,
+          databaseSessionBaseUrl: databaseSession.baseUrl,
+          configBaseUrl: config.baseUrl,
+        });
+
+        latestSession.value = null;
+        databaseSession = null;
+
+        break outer;
+      }
+
+      if (databaseSession.modelId !== config.modelId) {
+        console.warn(`Config model ID mismatch`, {
+          databaseSessionId: databaseSession.id,
+          databaseSessionModelId: databaseSession.modelId,
+          configModelId: config.modelId,
+        });
+
+        latestSession.value = null;
+        databaseSession = null;
+
+        break outer;
+      }
+    }
+
     const models = await api.v1.models.index(config.baseUrl);
     const model = models.find((model) => model.id === config.modelId);
     if (!model) {
       throw new Error(`Model ${config.modelId} not found`);
     }
 
-    return new RemoteLlmDriver(config, model.contextSize, jwt, sessionId);
+    return new RemoteLlmDriver(
+      config,
+      model.contextSize,
+      jwt,
+      latestSession,
+      databaseSession?.externalId,
+    );
   }
 
   readonly grammarLang = LlmGrammarLang.Regex;
@@ -42,7 +99,8 @@ export class RemoteLlmDriver implements BaseLlmDriver {
     readonly config: RemoteLlmDriverConfig,
     readonly contextSize: number,
     readonly jwt: Readonly<Ref<string | null>>,
-    readonly sessionId: Ref<string | null>,
+    private _latestSession: Ref<LatestSession | null>,
+    private _externalSessionId: string | undefined,
   ) {}
 
   compareConfig(other: RemoteLlmDriverConfig): boolean {
@@ -60,19 +118,16 @@ export class RemoteLlmDriver implements BaseLlmDriver {
     _decodeCallback?: (event: DecodeProgressEventPayload) => void,
     _completionCallback?: (event: CompletionProgressEventPayload) => void,
     abortSignal?: AbortSignal,
-  ): Promise<{
-    result: string;
-    totalTokens: number;
-    aborted: boolean;
-  }> {
+  ): Promise<CompletionResult> {
     if (!this.jwt.value) {
       throw new Error("JWT token not set");
     }
 
+    const startedAt = Date.now();
     try {
       this.busy.value = true;
 
-      const fetchTimeout = timeoutSignal(toMilliseconds({ minutes: 5 }));
+      const fetchTimeout = timeoutSignal(toMilliseconds({ minutes: 10 }));
       let signal;
       if (abortSignal) {
         signal = mergeAbortSignals(abortSignal, fetchTimeout);
@@ -83,7 +138,7 @@ export class RemoteLlmDriver implements BaseLlmDriver {
       const response = await api.v1.completions.create(
         this.config.baseUrl,
         this.jwt.value,
-        this.sessionId.value ?? undefined,
+        this._externalSessionId,
         {
           model: this.config.modelId,
           prompt: prompt,
@@ -100,13 +155,79 @@ export class RemoteLlmDriver implements BaseLlmDriver {
         signal,
       );
 
-      this.sessionId.value = response.sessionId;
+      console.debug("Response", omit(response, ["output"]));
+
+      // If the response contains a new session ID, store it in the database.
+      // The old session seems to be invalidated after a while.
+      if (response.sessionId !== this._externalSessionId) {
+        this._latestSession.value = {
+          driver: "remote",
+          id: (
+            await d.db
+              .insert(d.llmRemoteSessions)
+              .values({
+                externalId: response.sessionId,
+                baseUrl: this.config.baseUrl,
+                modelId: this.config.modelId,
+                contextSize: this.contextSize,
+              })
+              .returning({
+                id: d.llmRemoteSessions.id,
+              })
+          )[0].id,
+        };
+      }
+
+      this._externalSessionId = response.sessionId;
+      if (!(this._latestSession.value?.driver === "remote")) {
+        throw new Error("[BUG] Latest session value is not set");
+      }
+
+      const completion = (
+        await d.db
+          .insert(d.llmCompletions)
+          .values({
+            remoteSessionId: this._latestSession.value.id,
+            options: inferenceOptions,
+            input: prompt,
+            inputLength: response.usage.promptTokens,
+            output: response.output,
+            outputLength: response.usage.completionTokens,
+            error: abortSignal?.aborted ? "Aborted" : undefined,
+            delayTime: response.usage.delayTime,
+            executionTime: response.usage.executionTime,
+            realTime: Date.now() - startedAt,
+          })
+          .returning()
+      )[0];
 
       return {
-        result: response.output,
-        totalTokens: response.usage.totalTokens,
+        completion: {
+          id: completion.id,
+          input: completion.input,
+          inputLength: response.usage.promptTokens,
+          output: response.output,
+          outputLength: response.usage.completionTokens,
+        },
         aborted: abortSignal?.aborted ?? false,
       };
+    } catch (e: any) {
+      if (this._latestSession.value?.driver === "remote") {
+        await d.db.insert(d.llmCompletions).values({
+          remoteSessionId: this._latestSession.value.id,
+          options: inferenceOptions,
+          input: prompt,
+          error: e.message,
+          realTime: Date.now() - startedAt,
+        });
+      } else {
+        console.warn(
+          `Latest session is not remote, can't save errornous completion`,
+          { latestSession: this._latestSession.value },
+        );
+      }
+
+      throw e;
     } finally {
       this.busy.value = false;
     }