+++ packages/client/src/lib/storage.ts
@@ -2,7 +2,9 @@ import { useLocalStorage } from "@vueuse/core";
 
 export const remoteServerJwt = useLocalStorage<string | null>(
   "remoteServerJwt",
-  null,
+
+  // ADHOC: Use the Vite environment variable for the API JWT.
+  import.meta.env.VITE_API_JWT,
 );
 
 export * as llm from "./storage/llm";